Moved logic to pinMan subdirectory
